+++ src/result_processor.c
@@ -4,6 +4,7 @@
 #include <util/minmax_heap.h>
 #include "ext/default.h"
 #include "rmutil/rm_assert.h"
+#include "util/timeout.h"
 
 /*******************************************************************************************************************
  *  General Result Processor Helper functions
@@ -68,11 +69,8 @@ static int rpidxNext(ResultProcessor *base, SearchResult *res) {
   RPIndexIterator *self = (RPIndexIterator *)base;
   IndexIterator *it = self->iiter;
 
-  if (++self->timeoutLimiter == 100) {
-    self->timeoutLimiter = 0;
-    if (TimedOut(self->timeout) == RS_RESULT_TIMEDOUT) {
-      return RS_RESULT_TIMEDOUT;
-    }
+  if (TimedOut(self->timeout, &self->timeoutLimiter)) {
+    return RS_RESULT_TIMEDOUT;
   }
 
   // No root filter - the query has 0 results
